+++ src/platform_impl/web/event_loop/window_target.rs
@@ -18,7 +18,7 @@ use super::{
 };
 use crate::event::{
     DeviceEvent, DeviceId as RootDeviceId, ElementState, Event, KeyEvent, RawKeyEvent, Touch,
-    TouchPhase, WindowEvent,
+    TouchPhase, WindowEvent, WindowState,
 };
 use crate::event_loop::DeviceEvents;
 use crate::keyboard::ModifiersState;
@@ -714,10 +714,39 @@ impl<T> EventLoopWindowTarget<T> {
                         canvas.set_old_size(new_size);
                         runner.send_event(Event::WindowEvent {
                             window_id: RootWindowId(id),
-                            event: WindowEvent::Resized(new_size),
+                            event: WindowEvent::Configured {
+                                size: new_size,
+                                state: if backend::is_fullscreen(canvas.document(), canvas.raw()) {
+                                    WindowState::FULLSCREEN
+                                } else {
+                                    WindowState::empty()
+                                },
+                            },
                         });
                         runner.request_redraw(RootWindowId(id));
                     }
+                }
+            },
+            {
+                let runner = self.runner.clone();
+                let canvas = canvas_clone.clone();
+
+                move |new_size, is_fullscreen| {
+                    let canvas = RefCell::borrow(&canvas);
+                    canvas.set_current_size(new_size);
+                    canvas.set_old_size(new_size);
+                    runner.send_event(Event::WindowEvent {
+                        window_id: RootWindowId(id),
+                        event: WindowEvent::Configured {
+                            size: new_size,
+                            state: if is_fullscreen {
+                                WindowState::FULLSCREEN
+                            } else {
+                                WindowState::empty()
+                            },
+                        },
+                    });
+                    runner.request_redraw(RootWindowId(id));
                 }
             },
         );